+++ bee-network/src/peer/list.rs
@@ -5,10 +5,10 @@
 
 use super::{
     error::Error,
-    meta::{PeerInfo, PeerRelation},
+    info::{PeerInfo, PeerRelation},
 };
 
-use crate::{alias, config::Peer, init::global::max_unknown_peers, swarm::protocols::gossip::io::GossipSender};
+use crate::{alias, config::Peer, init::global, swarm::protocols::iota_gossip::GossipSender};
 
 use hashbrown::{HashMap, HashSet};
 use libp2p::{Multiaddr, PeerId};
@@ -158,13 +158,13 @@ impl PeerList {
             .map(|(info, state)| predicate(info, state))
     }
 
-    pub fn filter<'a, P: 'a>(&'a self, predicate: P) -> impl Iterator<Item = (PeerId, String)> + '_
+    pub fn filter_info<'a, P: 'a>(&'a self, predicate: P) -> impl Iterator<Item = (PeerId, PeerInfo)> + '_
     where
         P: Fn(&PeerInfo, &PeerState) -> bool,
     {
         self.peers.iter().filter_map(move |(peer_id, (info, state))| {
             if predicate(info, state) {
-                Some((*peer_id, info.alias.clone()))
+                Some((*peer_id, info.clone()))
             } else {
                 None
             }
@@ -236,11 +236,11 @@ impl PeerList {
         }
     }
 
-    pub fn unban_address(&mut self, address: &Multiaddr) -> Result<(), Error> {
-        if self.banned_addrs.remove(address) {
+    pub fn unban_address(&mut self, addr: &Multiaddr) -> Result<(), Error> {
+        if self.banned_addrs.remove(addr) {
             Ok(())
         } else {
-            Err(Error::AddressIsUnbanned(address.clone()))
+            Err(Error::AddressIsUnbanned(addr.clone()))
         }
     }
 
@@ -248,34 +248,35 @@ impl PeerList {
         self.banned_peers.contains(peer_id)
     }
 
-    pub fn is_addr_banned(&self, address: &Multiaddr) -> bool {
-        self.banned_addrs.contains(address)
+    pub fn is_addr_banned(&self, addr: &Multiaddr) -> bool {
+        self.banned_addrs.contains(addr)
     }
 
-    pub fn accepts_incoming_peer(&self, peer_id: &PeerId, peer_info: &PeerInfo) -> Result<(), Error> {
+    pub fn accepts_incoming_peer(&self, peer_id: &PeerId, peer_addr: &Multiaddr) -> Result<(), Error> {
+        // Checks performed are:
+        // - Deny ourself as peer.
+        // - Deny one of our own addresses.
+        // - Deny banned peers.
+        // - Deny banned addresses.
+        // - Deny already connected peers.
+        // - Deny more than the configured unknown peers.
         if peer_id == &self.local_id {
-            // Deny ourself as peer.
             Err(Error::PeerIsLocal(*peer_id))
-        } else if self.local_addrs.contains(&peer_info.address) {
-            // Deny one of our own addresses.
-            Err(Error::AddressIsLocal(peer_info.address.clone()))
+        } else if self.local_addrs.contains(&peer_addr) {
+            Err(Error::AddressIsLocal(peer_addr.clone()))
         } else if self.banned_peers.contains(peer_id) {
-            // Deny banned peers.
             Err(Error::PeerIsBanned(*peer_id))
-        } else if self.banned_addrs.contains(&peer_info.address) {
-            // Deny banned addresses.
-            Err(Error::AddressIsBanned(peer_info.address.clone()))
+        } else if self.banned_addrs.contains(&peer_addr) {
+            Err(Error::AddressIsBanned(peer_addr.clone()))
         } else if self
             .satisfies(peer_id, |_, state| state.is_connected())
             .unwrap_or(false)
         {
-            // Deny already connected peers.
             Err(Error::PeerIsConnected(*peer_id))
-        } else if peer_info.relation.is_unknown()
-            && self.filter_count(|info, _| info.relation.is_unknown()) >= max_unknown_peers()
+        } else if !self.contains(&peer_id)
+            && self.filter_count(|info, _| info.relation.is_unknown()) >= global::max_unknown_peers()
         {
-            // Deny more than configured unknown peers.
-            Err(Error::ExceedsUnknownPeerLimit(max_unknown_peers()))
+            Err(Error::ExceedsUnknownPeerLimit(global::max_unknown_peers()))
         } else {
             // All checks passed! Accept that peer.
             Ok(())
@@ -283,36 +284,36 @@ impl PeerList {
     }
 
     pub fn allows_dialing_peer(&self, peer_id: &PeerId) -> Result<(), Error> {
+        // Checks performed are:
+        // - Deny dialing ourself as peer.
+        // - Deny dialing a peer that has not been added first. TODO: check if we might want to allow this!
+        // - Deny dialing a banned peer.
+        // - Deny dialing an already connected peer.
+        // - Deny dialing a local address.
+        // - Deny dialing a banned address.
+        // - Deny dialing more than configured unkown peers.
         if peer_id == &self.local_id {
-            // Deny dialing ourself as peer.
             Err(Error::PeerIsLocal(*peer_id))
         } else if !self.contains(peer_id) {
-            // Deny dialing a peer that has not been added first.
-            // TODO: we might want to allow this.
             Err(Error::PeerNotPresent(*peer_id))
         } else if self.banned_peers.contains(peer_id) {
-            // Deny dialing a banned peer.
             Err(Error::PeerIsBanned(*peer_id))
         } else if self
             .satisfies(peer_id, |_, state| state.is_connected())
             .unwrap_or(false)
         {
-            // Deny dialing an already connected peer.
             Err(Error::PeerIsConnected(*peer_id))
         } else {
             let (peer_info, _) = self.peers.get(peer_id).unwrap();
 
             if self.local_addrs.contains(&peer_info.address) {
-                // Deny dialing a local address.
                 Err(Error::AddressIsLocal(peer_info.address.clone()))
             } else if self.banned_addrs.contains(&peer_info.address) {
-                // Deny dialing a banned address.
                 Err(Error::AddressIsBanned(peer_info.address.clone()))
             } else if peer_info.relation.is_unknown()
-                && self.filter_count(|info, _| info.relation.is_unknown()) >= max_unknown_peers()
+                && self.filter_count(|info, _| info.relation.is_unknown()) >= global::max_unknown_peers()
             {
-                // Deny dialing more than configured unkown peers.
-                Err(Error::ExceedsUnknownPeerLimit(max_unknown_peers()))
+                Err(Error::ExceedsUnknownPeerLimit(global::max_unknown_peers()))
             } else {
                 // All checks passed! Allow dialing that peer.
                 Ok(())
@@ -321,14 +322,15 @@ impl PeerList {
     }
 
     pub fn allows_dialing_addr(&self, addr: &Multiaddr) -> Result<(), Error> {
+        // Checks performed are:
+        // - Deny dialing a local address.
+        // - Deny dialing a banned address.
+        // - Deny dialing an already connected peer (with that address).
         if self.local_addrs.contains(addr) {
-            // Deny dialing a local address.
             Err(Error::AddressIsLocal(addr.clone()))
         } else if self.banned_addrs.contains(addr) {
-            // Deny dialing a banned address.
             Err(Error::AddressIsBanned(addr.clone()))
         } else if let Some(peer_id) = self.find_peer_if_connected(addr) {
-            // Deny dialing an already connected peer (with that address).
             Err(Error::PeerIsConnected(peer_id))
         } else {
             // All checks passed! Allow dialing that address.
@@ -337,8 +339,7 @@ impl PeerList {
     }
 
     fn find_peer_if_connected(&self, addr: &Multiaddr) -> Option<PeerId> {
-        #[allow(clippy::filter_next)]
-        self.filter(|info, state| state.is_connected() && info.address == *addr)
+        self.filter_info(|info, state| state.is_connected() && info.address == *addr)
             .next()
             .map(|(peer_id, _)| peer_id)
     }
@@ -396,7 +397,7 @@ mod tests {
         let pl = PeerList::new(local_id);
 
         assert!(matches!(
-            pl.accepts_incoming_peer(&local_id, &gen_constant_peer_info()),
+            pl.accepts_incoming_peer(&local_id, &gen_constant_peer_info().address),
             Err(Error::PeerIsLocal(_))
         ));
     }
@@ -410,7 +411,7 @@ mod tests {
         let mut pl = PeerList::new(local_id);
 
         pl.insert_peer(peer_id, peer_info.clone()).unwrap();
-        pl.accepts_incoming_peer(&peer_id, &peer_info).unwrap();
+        pl.accepts_incoming_peer(&peer_id, &peer_info.address).unwrap();
     }
 
     #[test]
@@ -431,9 +432,7 @@ mod tests {
         assert_eq!(0, pl.len());
     }
 
-    // =======================================================
-    // helpers
-    // =======================================================
+    // ===== helpers =====
 
     pub fn gen_constant_peer_id() -> PeerId {
         "12D3KooWJWEKvSFbben74C7H4YtKjhPMTDxd7gP7zxWSUEeF27st".parse().unwrap()
@@ -504,7 +503,7 @@ impl PeerState {
 #[cfg(test)]
 mod peerstate_tests {
     use super::*;
-    use crate::swarm::protocols::gossip::io::gossip_channel;
+    use crate::swarm::protocols::iota_gossip::channel;
 
     #[test]
     fn new_peer_state() {
@@ -516,7 +515,7 @@ mod peerstate_tests {
     #[test]
     fn peer_state_change() {
         let mut peerstate = PeerState::Disconnected;
-        let (tx, _rx) = gossip_channel();
+        let (tx, _rx) = channel();
 
         peerstate.set_connected(tx);
         assert!(peerstate.is_connected());