+++ bee-network/src/peer/mod.rs
@@ -2,5 +2,5 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod error;
+pub mod info;
 pub mod list;
-pub mod meta;