+++ bee-network/src/service/host.rs
@@ -12,19 +12,26 @@ use crate::{
     init::global::reconnect_interval_secs,
     peer::{
         error::Error as PeerError,
+        info::{PeerInfo, PeerRelation},
         list::PeerListWrapper as PeerList,
-        meta::{PeerInfo, PeerRelation},
     },
+    swarm::protocols::iota_gossip,
 };
 
 use bee_runtime::shutdown_stream::ShutdownStream;
 
-use futures::{channel::oneshot, StreamExt};
+use futures::{
+    channel::oneshot,
+    io::{BufReader, BufWriter},
+    AsyncReadExt, StreamExt,
+};
 use libp2p::{identity, Multiaddr, PeerId};
 use log::*;
 use rand::Rng;
 use tokio::time::{self, Duration, Instant};
 use tokio_stream::wrappers::{IntervalStream, UnboundedReceiverStream};
+
+const MAX_PEER_STATE_CHECKER_DELAY_MILLIS: u64 = 2000;
 
 pub struct ServiceHostConfig {
     pub local_keys: identity::Keypair,
@@ -46,6 +53,8 @@ pub struct Receivers {
 }
 
 type Shutdown = oneshot::Receiver<()>;
+
+const IO_BUFFER_LEN: usize = 32 * 1024;
 
 pub mod integrated {
     use super::*;
@@ -173,10 +182,10 @@ async fn command_processor(shutdown: Shutdown, commands: CommandReceiver, sender
 async fn event_processor(shutdown: Shutdown, events: InternalEventReceiver, senders: Senders, peerlist: PeerList) {
     debug!("Event processor running.");
 
-    let mut int_events = ShutdownStream::new(shutdown, UnboundedReceiverStream::new(events));
+    let mut internal_events = ShutdownStream::new(shutdown, UnboundedReceiverStream::new(events));
 
-    while let Some(int_event) = int_events.next().await {
-        if let Err(e) = process_internal_event(int_event, &senders, &peerlist).await {
+    while let Some(internal_event) = internal_events.next().await {
+        if let Err(e) = process_internal_event(internal_event, &senders, &peerlist).await {
             error!("Error processing internal event. Cause: {}", e);
             continue;
         }
@@ -195,7 +204,7 @@ async fn peerstate_checker(shutdown: Shutdown, senders: Senders, peerlist: PeerL
     // We want to reduce the overhead of simultaneous mutual dialing even if several nodes are started at the same time
     // (by script for example). We do this here by adding a small random delay to when this task will be executing
     // regular peer state checks.
-    let delay = Duration::from_millis(rand::thread_rng().gen_range(0u64..1000));
+    let delay = Duration::from_millis(rand::thread_rng().gen_range(0u64..MAX_PEER_STATE_CHECKER_DELAY_MILLIS));
     let start = Instant::now() + delay;
 
     // The (currently) constant interval at which peer state checks happen.
@@ -208,8 +217,18 @@ async fn peerstate_checker(shutdown: Shutdown, senders: Senders, peerlist: PeerL
     while interval.next().await.is_some() {
         let peerlist = peerlist.0.read().await;
 
-        for (peer_id, alias) in peerlist.filter(|info, state| info.relation.is_known() && state.is_disconnected()) {
-            info!("Trying to reconnect to: {} ({}).", alias, alias!(peer_id));
+        let num_known = peerlist.filter_count(|info, _| info.relation.is_known());
+        let num_connected_known = peerlist.filter_count(|info, state| info.relation.is_known() && state.is_connected());
+        let num_connected_unknown =
+            peerlist.filter_count(|info, state| info.relation.is_unknown() && state.is_connected());
+
+        info!(
+            "Connected peers: known {}/{} unknown {}.",
+            num_connected_known, num_known, num_connected_unknown,
+        );
+
+        for (peer_id, info) in peerlist.filter_info(|info, state| info.relation.is_known() && state.is_disconnected()) {
+            info!("Trying to connect to: {} ({}).", info.alias, alias!(peer_id));
 
             // Ignore if the command fails. We can always retry the next time.
             let _ = internal_commands.send(Command::DialPeer { peer_id });
@@ -232,14 +251,6 @@ async fn process_command(command: Command, senders: &Senders, peerlist: &PeerLis
             let alias = alias.unwrap_or_else(|| alias!(peer_id).to_string());
 
             add_peer(peer_id, multiaddr, alias, relation, senders, peerlist).await?;
-
-            // Automatically connect to "known" peers.
-            if relation.is_known() {
-                senders
-                    .internal_commands
-                    .send(Command::DialPeer { peer_id })
-                    .map_err(|_| Error::SendingCommandFailed)?;
-            }
         }
 
         Command::BanAddress { address } => {
@@ -312,8 +323,12 @@ async fn process_command(command: Command, senders: &Senders, peerlist: &PeerLis
     Ok(())
 }
 
-async fn process_internal_event(int_event: InternalEvent, senders: &Senders, peerlist: &PeerList) -> Result<(), Error> {
-    match int_event {
+async fn process_internal_event(
+    internal_event: InternalEvent,
+    senders: &Senders,
+    peerlist: &PeerList,
+) -> Result<(), Error> {
+    match internal_event {
         InternalEvent::AddressBound { address } => {
             senders
                 .events
@@ -342,59 +357,86 @@ async fn process_internal_event(int_event: InternalEvent, senders: &Senders, pee
         InternalEvent::ProtocolEstablished {
             peer_id,
             peer_addr,
-            conn_info,
-            gossip_in,
-            gossip_out,
+            origin,
+            substream,
         } => {
             let mut peerlist = peerlist.0.write().await;
+            let mut peer_added = false;
 
-            // In case the peer doesn't exist yet, we create a `PeerInfo` for that peer on-the-fly.
-            if !peerlist.contains(&peer_id) {
-                let peer_info = PeerInfo {
-                    address: peer_addr,
-                    alias: alias!(peer_id).to_string(),
-                    relation: PeerRelation::Unknown,
-                };
+            // NOTE: It's a bit unfortunate that atm there seems to be no way to inject custom criteria to prevent
+            // protocol negotiation. So we have to run the checks whether we want to allow that peer - and spend
+            // resources on it - here.
 
-                peerlist
-                    .insert_peer(peer_id, peer_info.clone())
-                    .map_err(|(_, _, e)| e)?;
+            let accepted = peerlist.accepts_incoming_peer(&peer_id, &peer_addr);
+
+            if accepted.is_ok() {
+                // If the peer doesn't exist yet - but is accepted as an "unknown" peer, we insert it now.
+                if !peerlist.contains(&peer_id) {
+                    let peer_info = PeerInfo {
+                        address: peer_addr,
+                        alias: alias!(peer_id).to_string(),
+                        relation: PeerRelation::Unknown,
+                    };
+                    peerlist.insert_peer(peer_id, peer_info).map_err(|(_, _, e)| e)?;
+                    peer_added = true;
+                }
+
+                // Panic:
+                // We made sure, that the peer id exists in the above if-branch, hence, unwrapping is fine.
+                let peer_info = peerlist.info(&peer_id).unwrap();
+
+                // Spin up separate buffered reader and writer to efficiently process the gossip with that peer.
+                let (r, w) = substream.split();
+
+                let reader = BufReader::with_capacity(IO_BUFFER_LEN, r);
+                let writer = BufWriter::with_capacity(IO_BUFFER_LEN, w);
+
+                let (incoming_tx, incoming_rx) = iota_gossip::channel();
+                let (outgoing_tx, outgoing_rx) = iota_gossip::channel();
+
+                iota_gossip::start_incoming_processor(peer_id, reader, incoming_tx, senders.internal_events.clone());
+                iota_gossip::start_outgoing_processor(peer_id, writer, outgoing_rx, senders.internal_events.clone());
+
+                // We store a clone of the gossip send channel in order to send a shutdown signal.
+                let _ = peerlist.update_state(&peer_id, |state| state.set_connected(outgoing_tx.clone()));
+
+                // We no longer need to hold the lock.
+                drop(peerlist);
+
+                // We only want to fire events when no longer holding the lock to the peerlist to make this code more
+                // resilient against different channel implementations.
+
+                if peer_added {
+                    senders
+                        .events
+                        .send(Event::PeerAdded {
+                            peer_id,
+                            info: peer_info.clone(),
+                        })
+                        .map_err(|_| Error::SendingEventFailed)?;
+                }
+
+                info!(
+                    "Established ({}) protocol with {} ({}).",
+                    origin,
+                    peer_info.alias,
+                    alias!(peer_id)
+                );
 
                 senders
                     .events
-                    .send(Event::PeerAdded {
+                    .send(Event::PeerConnected {
                         peer_id,
                         info: peer_info,
+                        gossip_in: incoming_rx,
+                        gossip_out: outgoing_tx,
                     })
                     .map_err(|_| Error::SendingEventFailed)?;
+            } else {
+                // Panic:
+                // This branch handles the error case, so unwrapping it is fine.
+                debug!("{}", accepted.unwrap_err());
             }
-
-            // Panic:
-            // We made sure, that the peer id exists in the above if-branch, hence, unwrapping is fine.
-            let peer_info = peerlist.info(&peer_id).unwrap();
-
-            // We store a clone of the gossip send channel in order to send a shutdown signal.
-            let _ = peerlist.update_state(&peer_id, |state| state.set_connected(gossip_out.clone()));
-
-            // We no longer need to hold the lock.
-            drop(peerlist);
-
-            info!(
-                "Established ({}) protocol with {} ({}).",
-                conn_info.origin,
-                peer_info.alias,
-                alias!(peer_id)
-            );
-
-            senders
-                .events
-                .send(Event::PeerConnected {
-                    peer_id,
-                    info: peer_info,
-                    gossip_in,
-                    gossip_out,
-                })
-                .map_err(|_| Error::SendingEventFailed)?;
         }
     }