+++ bee-network/src/swarm/protocols/gossip/mod.rs
@@ -1,8 +1,0 @@
-// Copyright 2020-2021 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod behavior;
-pub mod event;
-pub mod handler;
-pub mod io;
-pub mod upgrade;